optimize out recursive calls
This is an attempt at avoiding overflowing the stack with too much
recursion.

This patch implements CPS + a trampoline. Every Finite.nth getter is
rewritten in CPS, but instead of calling the next continuation it
returns it (bound to its `this` and arguments). The outer `nth` function
is the trampoline which calls the continuation one after the other and
reads the final result.

All these function objects are probably causing too much allocation.
It's a future excercise to try to measure that and optimize it away
(defunctionalization?).

The types for Continuation and BoundContinuation are probably also not
perfect, but they were good enough to help me to reason about this.